Fix: Fixed unprotected initate user mutation. Reduced point of failur…

+++ frontend/convex/keysMutations.ts
@@ -2,15 +2,21 @@ import { v } from "convex/values";
 import { internalMutation, internalQuery, query } from "./_generated/server";
 
 export const getEncryptionKeys = internalQuery({
-  args: {
-    user_id: v.string(),
-  },
-  handler: async (ctx, args) => {
-    const { user_id } = args;
+  handler: async (ctx) => {
+    // const { user_id } = args;
+    const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authorized");
+
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_ExternalId", (q) => q.eq("externalId", identity.subject))
+      .unique();
+
+    if (user === null) throw new Error("User not found");
 
     return await ctx.db
       .query("userEncryptionKeys")
-      .withIndex("by_user", (q) => q.eq("user_id", user_id))
+      .withIndex("by_userId", (q) => q.eq("userId", user._id))
       .first();
   },
 });
@@ -20,7 +26,7 @@ export const ensureEncryptionKeys = internalMutation({
     entropy: v.string(),
     salt: v.string(),
     version: v.string(),
-    kdf_name: v.literal('scrypt'), // add argon2 later
+    kdf_name: v.literal("scrypt"), // add argon2 later
     params: v.object({
       N: v.number(),
       r: v.number(),
@@ -28,14 +34,21 @@ export const ensureEncryptionKeys = internalMutation({
     }),
   },
   handler: async (ctx, args) => {
-    const idendity = await ctx.auth.getUserIdentity();
-    if (!idendity) throw new Error("Not authenticated");
+    const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authenticated");
+
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_ExternalId", (q) => q.eq("externalId", identity.subject))
+      .unique();
+
+    if (user === null) throw new Error("User not found");
 
     const { entropy, salt, version, kdf_name, params } = args;
 
     const encryptionKey = await ctx.db
       .query("userEncryptionKeys")
-      .withIndex("by_user", (q) => q.eq("user_id", idendity.subject))
+      .withIndex("by_userId", (q) => q.eq("userId", user._id))
       .first();
 
     if (encryptionKey) {
@@ -43,7 +56,7 @@ export const ensureEncryptionKeys = internalMutation({
     }
 
     const insertedKey = await ctx.db.insert("userEncryptionKeys", {
-      user_id: idendity.subject,
+      userId: user._id,
       entropy,
       salt,
       version,
@@ -54,25 +67,32 @@ export const ensureEncryptionKeys = internalMutation({
     const insertedEncryptionKey = await ctx.db.get(insertedKey);
 
     return insertedEncryptionKey;
-
-  }
-})
-
+  },
+});
 
 export const ensureEncryptedApiKeys = internalMutation({
   args: {
-    user_id: v.string(),
     provider: v.string(),
     encryptedApiKey: v.string(),
     derivedAt: v.number(),
   },
   handler: async (ctx, args) => {
-    const { user_id, provider, encryptedApiKey, derivedAt } = args;
+    const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authenticated");
+
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_ExternalId", (q) => q.eq("externalId", identity.subject))
+      .unique();
+
+    if (user === null) throw new Error("User not found");
+
+    const { provider, encryptedApiKey, derivedAt } = args;
 
     const existing = await ctx.db
       .query("userApiKeys")
       .withIndex("by_user_provider", (q) =>
-        q.eq("user_id", user_id).eq("provider", provider)
+        q.eq("userId", user._id).eq("provider", provider),
       )
       .first();
 
@@ -80,12 +100,12 @@ export const ensureEncryptedApiKeys = internalMutation({
       await ctx.db.patch(existing._id, {
         encryptedApiKey,
         derivedAt,
-      })
+      });
       return await ctx.db.get(existing._id);
     }
 
     const encryptApiKeyId = await ctx.db.insert("userApiKeys", {
-      user_id: user_id,
+      userId: user._id,
       provider: provider,
       encryptedApiKey: encryptedApiKey,
       derivedAt,
@@ -99,16 +119,25 @@ export const ensureEncryptedApiKeys = internalMutation({
 
 export const getEncryptedApiKey = internalQuery({
   args: {
-    user_id: v.string(),
     provider: v.string(),
   },
   handler: async (ctx, args) => {
-    const { user_id, provider } = args;
+    const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authenticated");
+
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_ExternalId", (q) => q.eq("externalId", identity.subject))
+      .unique();
+
+    if (user === null) throw new Error("User not found");
+
+    const { provider } = args;
 
     return await ctx.db
       .query("userApiKeys")
       .withIndex("by_user_provider", (q) =>
-        q.eq("user_id", user_id).eq("provider", provider)
+        q.eq("userId", user._id).eq("provider", provider),
       )
       .first();
   },
@@ -117,12 +146,18 @@ export const getEncryptedApiKey = internalQuery({
 export const getAllApiKeys = query({
   handler: async (ctx) => {
     const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authenticated");
 
-    if (!identity) throw new Error("User not authenticated");
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_ExternalId", (q) => q.eq("externalId", identity.subject))
+      .unique();
+
+    if (user === null) throw new Error("User not found");
 
     return await ctx.db
       .query("userApiKeys")
-      .withIndex("by_user", (q) => q.eq("user_id", identity.subject))
+      .withIndex("by_userId", (q) => q.eq("userId", user._id))
       .collect();
   },
 });
@@ -132,6 +167,9 @@ export const deleteApiKey = internalMutation({
     keyId: v.id("userApiKeys"),
   },
   handler: async (ctx, args) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if (identity === null) throw new Error("Not authenticated");
+
     const { keyId } = args;
 
     await ctx.db.delete(keyId);
